Create hard links of reboot and poweroff to shutdown
