+++ elkscmd/Applications
@@ -87,13 +87,13 @@ sys_utils/chmem                 :sysutil            :720k
 sys_utils/kill                  :sysutil        :360c :720k
 sys_utils/ps            :sash   :sysutil        :360k           :128k       :nocomp
 sys_utils/ps :::uptime          :sysutil                        :128k   :1440k
-sys_utils/reboot                :sysutil        :360k           :128k
 sys_utils/makeboot              :sysutil        :360k
 sys_utils/man                   :sysutil                :1200k
 sys_utils/meminfo       :sash   :sysutil        :360k           :128k
 sys_utils/mouse                 :sysutil                 :1200c     :1440k
 sys_utils/passwd                :sysutil                :1200k
-sys_utils/poweroff              :sysutil            :720k
+sys_utils/shutdown :::reboot    :sysutil        :360k           :128k
+sys_utils/shutdown :::poweroff  :sysutil            :720k
 sys_utils/sercat                :sysutil                            :1400c
 sys_utils/console               :sysutil                :1200k
 #sys_utils/who                  :sysutil                :1200k
